hosted/src/content-scout/requirements.txt
# Agent Framework + Foundry hosting
agent-framework>=1.1.0
agent-framework-foundry-hosting

# Azure auth + optional KV/Storage
azure-identity>=1.17.0
azure-keyvault-secrets>=4.8.0
azure-storage-blob>=12.22.0

# Scanner HTTP client
httpx>=0.27.0

# Env loader (local dev)
python-dotenv>=1.0.0
